+++ README-HBBFT.md
@@ -8,9 +8,10 @@ Tolerant consensus algorithm](https://github.com/poanetwork/hbbft).
 ### Running a Test Peer
 
 1. `git clone https://github.com/poanetwork/parity-ethereum`
-2. `cd parity-ethereum`
-3. `git checkout hbbft`
-3. `./indica-node 0`
+1. `cd parity-ethereum`
+1. `git checkout hbbft`
+1. `./setup-indica clear`
+1. `./indica-node 0`
 
 #### Additional Peers
 
@@ -27,6 +28,7 @@ Each peer will generate a number of random transactions at regular intervals,
 process them accordingly, and output complete batches. If your terminal is
 spammed with batch outputs, consensus is working.
 
+<!--
 ### Signer Nodes
 
 Running a node with an associated engine signer account allows the node to
@@ -37,7 +39,7 @@ requires a bit of set up to add the associated test accounts.
 NOTE: This information may be out of date due to ongoing changes to the
 engine.
 
-1. Run `setup-indica-signers clear`. This will clear all existing blockchain
+1. Run `setup-indica clear`. This will clear all existing blockchain
    data.
 2. Start all 3 nodes as described above.
 3. Run `setup-indica-signers` (without `clear` argument).
@@ -52,9 +54,9 @@ curl --data '{"jsonrpc":"2.0","method":"personal_sendTransaction","params":[{"fr
 curl --data '{"jsonrpc":"2.0","method":"eth_getBalance","params":["0x00Bd138aBD70e2F00903268F3Db08f2D25677C9e", "latest"],"id":1}' -H "Content-Type: application/json" -X POST localhost:8501
 ```
    The balance should be `0x1bc16d674ec90000`.
-
+-->
 
 
 ### Status
 
-Experimental
+Alpha
