+++ ethcore/src/block.rs
@@ -508,7 +508,7 @@ impl LockedBlock {
 	/// NOTE: This does not check the validity of `seal` with the engine.
 	pub fn seal(self, engine: &EthEngine, seal: Vec<Bytes>) -> Result<SealedBlock, BlockError> {
 		let expected_seal_fields = engine.seal_fields(self.header());
-		println!("######## LOCKEDBLOCK::SEAL: expected_seal_fields: {}, seal: {:?}, seal.len(): {}",
+		debug!("######## LOCKEDBLOCK::SEAL: expected_seal_fields: {}, seal: {:?}, seal.len(): {}",
 			expected_seal_fields, seal, seal.len());
 		let mut s = self;
 		if seal.len() != expected_seal_fields {
