+++ ethcore/src/client/client.rs
@@ -2323,7 +2323,7 @@ impl ImportSealedBlock for Client {
 
 			let route = self.importer.commit_block(block, &header, encoded::Block::new(block_data), self);
 			trace!(target: "client", "Imported sealed block #{} ({})", number, h);
-			info!(target: "client", "####### Imported sealed block #{} ({})", number, h);
+			debug!(target: "client", "####### Imported sealed block #{} ({})", number, h);
 			self.state_db.write().sync_cache(&route.enacted, &route.retracted, false);
 			route
 		};
